+++ src/gusanos/level.cpp
@@ -413,7 +413,7 @@ Vec CMap::getSpawnLocation(CWormInputHandler* player)
 	//pos = Vec(rnd() * material->w, rnd()*material->h);
 
 	do {
-		pos = Vec(rnd() * material->w, rnd()*material->h);
+		pos = Vec((float)rnd() * material->w, (float)rnd()*material->h);
 	} while ( !getMaterial( static_cast<int>(pos.x), static_cast<int>(pos.y) ).worm_pass );
 
 	return pos;
