+++ src/gusanos/lua51/luaapi/context.cpp
@@ -519,7 +519,7 @@ void LuaContext::serialize(Net_BitStream& s, int i)
 			int tab = i < 0 ? i - 1 : i;
 			while(lua_next(m_State, tab) != 0)
 			{
-				if(!lua_isnumber(m_State, -2) || lua_tointeger(m_State, -2) >= idx)
+				if(!lua_isnumber(m_State, -2) || lua_tointeger(m_State, -2) >= (int)idx)
 				{
 					serialize(s, -2);
 					serialize(s, -1);
@@ -787,7 +787,7 @@ void LuaContext::serializeT(std::ostream& s, int i, int indent)
 			int tab = i < 0 ? i - 1 : i;
 			while(lua_next(m_State, tab) != 0)
 			{
-				if(!lua_isnumber(m_State, -2) || lua_tointeger(m_State, -2) >= idx)
+				if(!lua_isnumber(m_State, -2) || lua_tointeger(m_State, -2) >= (int)idx)
 				{
 					for(int j = 0; j < indent; ++j)
 						s.put('\t');
